+++ plugin/src/lib.rs
@@ -1,0 +1,2 @@
+#![deny(unsafe_code)]
+#![doc = include_str!("../README.md")]